# src/main/java/org/example/nman/model/NoteModel.java
package org.example.nman.model;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;

@Setter
@Getter
@NoArgsConstructor
@AllArgsConstructor
@Document(collection = "notes")
public class NoteModel {

    @Id
    private String id;
    private String title;
    private String content;
}